+++ CHANGES.md
@@ -7,6 +7,7 @@
  - 2017-12-12 Always require cookies for OTRS login - drop support for non-cookie sessions.
 
 #6.0.10 2018-??-??
+ - 2018-07-17 Fixed bug#[13722](https://bugs.otrs.org/show_bug.cgi?id=13722) - Translation of ticket notification header uses the system language instead of recipients preferences.
 
 #6.0.9 2018-07-24
  - 2018-07-16 Updated translations, thanks to all translators.